Moved psalm-language-server to the CLI command class
